fix: sandbox belt phase is f64 — the shared core's state, verbatim
The sandbox kept its own f32 BeltPhase, converted into SideState, then
ignored report.state.phase and re-advanced the f32 manually — adapter
drift from the shared core (codex purity audit #2). Now the resource
IS the core's f64 phase, committed back verbatim; view consumers wrap
in f64 and cast at the seam, mirroring the game view. Harness rows
still emit %.4f — schema unchanged.

Measured against the old f32 baseline (500-tick reversal slam):
max |Δphase| 0.5 mm, |Δhull| 0.7 mm — pure f32-accumulation noise.
Parity baseline re-recorded on the f64 output.

244 tests green, clippy zero.

+++ src/track_sandbox/harness.rs
@@ -250,7 +250,7 @@ pub(super) fn harness_record(
         pitch,
         arr([lin.0.x, lin.0.y, lin.0.z]),
         arr([belt.left, belt.right]),
-        arr([phase.get(Side::Left), phase.get(Side::Right)]),
+        arr([phase.get(Side::Left) as f32, phase.get(Side::Right) as f32]),
         total,
         wheel_json.join(","),
         contact_rows.join(","),
